[refactor] #53 - TagSection 분리 및 svg 파일 분리
